requirements.txt
fastapi
uvicorn[standard]
tensorflow     # or tensorflow-cpu if you prefer CPU-only
pillow
numpy
python-multipart
